wip: rename all annotations-related files to contain ee suffix
