+++ packages/cli/src/databases/entities/annotation-tag-entity.ee.ts
@@ -1,8 +1,8 @@
 import { Column, Entity, Index, ManyToMany, OneToMany } from '@n8n/typeorm';
 import { IsString, Length } from 'class-validator';
 
-import type { AnnotationTagMapping } from '@/databases/entities/annotation-tag-mapping';
-import type { ExecutionAnnotation } from '@/databases/entities/execution-annotation';
+import type { AnnotationTagMapping } from '@/databases/entities/annotation-tag-mapping.ee';
+import type { ExecutionAnnotation } from '@/databases/entities/execution-annotation.ee';
 
 import { WithTimestampsAndStringId } from './abstract-entity';
