+++ packages/cli/src/databases/entities/annotation-tag-mapping.ee.ts
@@ -1,7 +1,7 @@
 import { Entity, JoinColumn, ManyToOne, PrimaryColumn } from '@n8n/typeorm';
 
-import type { AnnotationTagEntity } from './annotation-tag-entity';
-import type { ExecutionAnnotation } from './execution-annotation';
+import type { AnnotationTagEntity } from './annotation-tag-entity.ee';
+import type { ExecutionAnnotation } from './execution-annotation.ee';
 
 /**
  * This entity represents the junction table between the execution annotations and the tags
