+++ packages/cli/src/databases/repositories/annotation-tag-mapping.repository.ee.ts
@@ -1,7 +1,7 @@
 import { DataSource, Repository } from '@n8n/typeorm';
 import { Service } from 'typedi';
 
-import { AnnotationTagMapping } from '@/databases/entities/annotation-tag-mapping';
+import { AnnotationTagMapping } from '@/databases/entities/annotation-tag-mapping.ee';
 
 @Service()
 export class AnnotationTagMappingRepository extends Repository<AnnotationTagMapping> {
