+++ packages/cli/src/databases/repositories/annotation-tag.repository.ee.ts
@@ -1,7 +1,7 @@
 import { DataSource, Repository } from '@n8n/typeorm';
 import { Service } from 'typedi';
 
-import { AnnotationTagEntity } from '@/databases/entities/annotation-tag-entity';
+import { AnnotationTagEntity } from '@/databases/entities/annotation-tag-entity.ee';
 
 @Service()
 export class AnnotationTagRepository extends Repository<AnnotationTagEntity> {
