+++ packages/cli/src/databases/repositories/execution-annotation.repository.ts
@@ -1,7 +1,7 @@
 import { DataSource, Repository } from '@n8n/typeorm';
 import { Service } from 'typedi';
 
-import { ExecutionAnnotation } from '@/databases/entities/execution-annotation';
+import { ExecutionAnnotation } from '@/databases/entities/execution-annotation.ee';
 
 @Service()
 export class ExecutionAnnotationRepository extends Repository<ExecutionAnnotation> {
