+++ packages/cli/src/services/annotation-tag.service.ee.ts
@@ -1,7 +1,7 @@
 import { Service } from 'typedi';
 
-import type { AnnotationTagEntity } from '@/databases/entities/annotation-tag-entity';
-import { AnnotationTagRepository } from '@/databases/repositories/annotation-tag.repository';
+import type { AnnotationTagEntity } from '@/databases/entities/annotation-tag-entity.ee';
+import { AnnotationTagRepository } from '@/databases/repositories/annotation-tag.repository.ee';
 import { validateEntity } from '@/generic-helpers';
 import type { IAnnotationTagDb, IAnnotationTagWithCountDb } from '@/interfaces';
