+++ packages/cli/test/integration/shared/utils/test-server.ts
@@ -140,7 +140,7 @@ export const setupTestServer = ({
 			for (const group of endpointGroups) {
 				switch (group) {
 					case 'annotationTags':
-						await import('@/controllers/annotation-tags.controller');
+						await import('@/controllers/annotation-tags.controller.ee');
 						break;
 
 					case 'credentials':
