+++ packages/editor-ui/src/components/AnnotationTagsContainer.ee.vue
