+++ packages/editor-ui/src/components/AnnotationTagsDropdown.ee.vue
