+++ packages/editor-ui/src/components/TagsManager/AnnotationTagsManager.ee.vue
