+++ packages/editor-ui/src/components/executions/workflow/WorkflowExecutionAnnotationPanel.ee.vue
@@ -2,7 +2,7 @@
 import { ref, computed } from 'vue';
 import type { AnnotationVote, ExecutionSummary } from 'n8n-workflow';
 import { useExecutionsStore } from '@/stores/executions.store';
-import AnnotationTagsDropdown from '@/components/AnnotationTagsDropdown.vue';
+import AnnotationTagsDropdown from '@/components/AnnotationTagsDropdown.ee.vue';
 import { createEventBus } from 'n8n-design-system';
 import VoteButtons from '@/components/executions/workflow/VoteButtons.vue';
 import { useToast } from '@/composables/useToast';
